+++ PrinterChangeNotification/NotifyOptionsType.cs
@@ -7,6 +7,6 @@ namespace PrinterChangeNotification
     public class NotifyOptionsType
     {
         public NOTIFY_TYPE Type;
-        public List<UInt32> Fields = new List<UInt32>();
+        public List<UInt16> Fields = new List<UInt16>();
     }
 }
